requirements.txt
pynamodb
python-dotenv
Faker
boto3